allow param values to include any string
