+++ core/zms/src/main/java/com/yahoo/athenz/zms/ZMSSchema.java
@@ -387,7 +387,7 @@ public class ZMSSchema {
 
         sb.structType("TemplateParam")
             .field("name", "SimpleName", false, "name of the parameter")
-            .field("value", "CompoundName", false, "value of the parameter");
+            .field("value", "String", false, "value of the parameter");
 
         sb.structType("DomainTemplate", "TemplateList")
             .comment("solution template(s) to be applied to a domain")
