+++ libs/go/zmscli/template.go
@@ -90,7 +90,7 @@ func (cli Zms) SetDomainTemplate(dn string, templateArgs []string) (*string, err
 		} else {
 			param := zms.TemplateParam{
 				Name:  zms.SimpleName(value[0:idx]),
-				Value: zms.CompoundName(value[idx+1:]),
+				Value: value[idx+1:],
 			}
 			templateParams = append(templateParams, &param)
 		}
