+++ servers/zms/src/test/java/com/yahoo/athenz/zms/ZMSImplTest.java
@@ -29,7 +29,6 @@ import com.yahoo.athenz.auth.Principal;
 import com.yahoo.athenz.auth.ServerPrivateKey;
 import com.yahoo.athenz.auth.impl.*;
 import com.yahoo.athenz.auth.token.PrincipalToken;
-import com.yahoo.athenz.auth.util.AthenzUtils;
 import com.yahoo.athenz.auth.util.Crypto;
 import com.yahoo.athenz.common.config.AuthzDetailsEntity;
 import com.yahoo.athenz.common.config.AuthzDetailsField;
@@ -69,7 +68,6 @@ import jakarta.ws.rs.core.Response;
 import org.hamcrest.CoreMatchers;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
-import org.testng.Assert;
 import org.testng.annotations.*;
 
 import java.io.*;
@@ -5046,11 +5044,11 @@ public class ZMSImplTest {
 
         zmsImpl.deleteTopLevelDomain(ctx, "PolicyGetDom1", auditRef);
     }
+
     @Test
     public void testGetPolicyWithAssertionConditions() {
 
         TestAuditLogger alogger = new TestAuditLogger();
-        List<String> aLogMsgs = alogger.getLogMsgList();
         ZMSImpl zmsImpl = zmsTestInitializer.getZmsImpl(alogger);
         RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
         final String auditRef = zmsTestInitializer.getAuditRef();
@@ -8755,7 +8753,7 @@ public class ZMSImplTest {
                     Crypto.loadPublicKey(publicKey), signature));
             DomainData ddata = sDomain.getDomain();
             assertNotNull(ddata.getPolicies());
-            assertTrue(ddata.getRoles() != null && ddata.getRoles().size() > 0);
+            assertTrue(ddata.getRoles() != null && !ddata.getRoles().isEmpty());
             assertNotNull(ddata.getServices());
         }
 
@@ -8777,7 +8775,7 @@ public class ZMSImplTest {
                     Crypto.loadPublicKey(publicKey), signature));
             DomainData ddata = sDomain.getDomain();
             assertNotNull(ddata.getPolicies());
-            assertTrue(ddata.getRoles() != null && ddata.getRoles().size() > 0);
+            assertTrue(ddata.getRoles() != null && !ddata.getRoles().isEmpty());
             assertNotNull(ddata.getServices());
         }
 
@@ -10820,15 +10818,11 @@ public class ZMSImplTest {
         } catch (ResourceException ignored) {
         }
 
-        // remove the second element and add another with invalid value
+        // value is string so it will be accepted
 
         params.remove(1);
         params.add(new TemplateParam().setName("param_name_valid").setValue("param_value_invalid(again)"));
-        try {
-            zmsImpl.validate(domainTemplate, "DomainTemplate", "testValidateDomainTemplate");
-            fail();
-        } catch (ResourceException ignored) {
-        }
+        zmsImpl.validate(domainTemplate, "DomainTemplate", "testValidateDomainTemplate");
     }
 
     @Test
@@ -16577,788 +16571,6 @@ public class ZMSImplTest {
     }
 
     @Test
-    public void testPutDomainTemplateInvalidTemplate() {
-
-        TestAuditLogger alogger = new TestAuditLogger();
-        ZMSImpl zmsImpl = zmsTestInitializer.getZmsImpl(alogger);
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-        String domainName = "templatelist-invalid";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        // first no templates
-
-        DomainTemplate templateList = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templateList.setTemplateNames(templates);
-
-        try {
-            zmsImpl.putDomainTemplate(ctx, domainName, auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(ex.getCode(), 400);
-            assertTrue(ex.getMessage().contains("No templates specified"));
-        }
-
-        // then invalid template
-
-        templates.add("test validate");
-        try {
-            zmsImpl.putDomainTemplate(ctx, domainName, auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(400, ex.getCode());
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateNotFoundTemplate() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        String domainName = "templatelist-invalid";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        DomainTemplate templateList = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templates.add("InvalidTemplate");
-        templateList.setTemplateNames(templates);
-        try {
-            zmsImpl.putDomainTemplate(ctx, domainName, auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(404, ex.getCode());
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateSingleTemplate() {
-
-        String domainName = "templatelist-single";
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        SubDomain domSysNetwork = zmsTestInitializer.createSubDomainObject("network", "sys", "Test Domain",
-                "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postSubDomain(ctx, "sys", auditRef, domSysNetwork);
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        DomainTemplate domTemplate = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templates.add("vipng");
-        domTemplate.setTemplateNames(templates);
-
-        zmsImpl.putDomainTemplate(ctx, domainName, auditRef, domTemplate);
-
-        // verify that our role collection includes the roles defined in template
-
-        List<String> names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(3, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-
-        Role role = zmsImpl.dbService.getRole(domainName, "vip_admin", false, false, false);
-        assertEquals(domainName + ":role.vip_admin", role.getName());
-        assertNull(role.getTrust());
-        assertTrue(role.getRoleMembers().isEmpty());
-
-        role = zmsImpl.dbService.getRole(domainName, "sys_network_super_vip_admin", false, false, false);
-        assertEquals(domainName + ":role.sys_network_super_vip_admin", role.getName());
-        assertEquals("sys.network", role.getTrust());
-
-        // verify that our policy collections includes the policies defined in the template
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(3, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-
-        Policy policy = zmsImpl.dbService.getPolicy(domainName, "vip_admin", null);
-        assertEquals(domainName + ":policy.vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        Assertion assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(domainName + ":role.vip_admin", assertion.getRole());
-        assertEquals(domainName + ":vip*", assertion.getResource());
-
-        policy = zmsImpl.dbService.getPolicy(domainName, "sys_network_super_vip_admin", null);
-        assertEquals(domainName + ":policy.sys_network_super_vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(domainName + ":role.sys_network_super_vip_admin", assertion.getRole());
-        assertEquals(domainName + ":vip*", assertion.getResource());
-
-        // delete an applied service template
-        //
-        String templateName = "vipng";
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName, auditRef);
-
-        // verify that our role collection does NOT include the roles defined in template
-
-        names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        zmsImpl.deleteSubDomain(ctx, "sys", "network", auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateMultipleTemplates() {
-
-        String domainName = "templatelist-multiple";
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        SubDomain domSysNetwork = zmsTestInitializer.createSubDomainObject("network", "sys", "Test Domain",
-                "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postSubDomain(ctx, "sys", auditRef, domSysNetwork);
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        DomainTemplate domTemplate = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templates.add("vipng");
-        templates.add("platforms");
-        templates.add("user_provisioning");
-        domTemplate.setTemplateNames(templates);
-
-        zmsImpl.putDomainTemplate(ctx, domainName, auditRef, domTemplate);
-
-        // verify that our role collection includes the roles defined in template
-
-        List<String> names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(7, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-        assertTrue(names.contains("platforms_deployer"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        Role role = zmsImpl.dbService.getRole(domainName, "openstack_readers", false, false, false);
-        assertEquals(domainName + ":role.openstack_readers", role.getName());
-        assertNull(role.getTrust());
-        assertEquals(2, role.getRoleMembers().size());
-
-        List<String> checkList = new ArrayList<>();
-        checkList.add("sys.builder");
-        checkList.add("sys.openstack");
-        zmsTestInitializer.checkRoleMember(checkList, role.getRoleMembers());
-
-        role = zmsImpl.dbService.getRole(domainName, "sys_network_super_vip_admin", false, false, false);
-        assertEquals(domainName + ":role.sys_network_super_vip_admin", role.getName());
-        assertEquals("sys.network", role.getTrust());
-
-        // verify that our policy collections includes the policies defined in the template
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(7, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-        assertTrue(names.contains("platforms_deploy"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        Policy policy = zmsImpl.dbService.getPolicy(domainName, "vip_admin", null);
-        assertEquals(domainName + ":policy.vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        Assertion assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(domainName + ":role.vip_admin", assertion.getRole());
-        assertEquals(domainName + ":vip*", assertion.getResource());
-
-        policy = zmsImpl.dbService.getPolicy(domainName, "sys_network_super_vip_admin", null);
-        assertEquals(domainName + ":policy.sys_network_super_vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(domainName + ":role.sys_network_super_vip_admin", assertion.getRole());
-        assertEquals(domainName + ":vip*", assertion.getResource());
-
-        // delete applied service template
-        //
-        String templateName = "vipng";
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName, auditRef);
-
-        // verify that our role collection does NOT include the vipng roles defined in template
-
-        names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(5, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("platforms_deployer"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(5, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("platforms_deploy"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        // delete applied service template
-        //
-        templateName = "platforms";
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName, auditRef);
-
-        // verify that our role collection does NOT include the platforms roles defined in template
-
-        names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(4, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(4, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        // delete last applied service template
-        //
-        templateName = "user_provisioning";
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName, auditRef);
-
-        // verify that our role collection does NOT include the user_provisioning roles defined in template
-
-        names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        zmsImpl.deleteSubDomain(ctx, "sys", "network", auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateExtInvalidTemplate() {
-
-        TestAuditLogger alogger = new TestAuditLogger();
-        ZMSImpl zmsImpl = zmsTestInitializer.getZmsImpl(alogger);
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        String domainName = "templatelist-invalid";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        // first no templates
-
-        DomainTemplate templateList = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templateList.setTemplateNames(templates);
-
-        try {
-            zmsImpl.putDomainTemplate(ctx, domainName, auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(ex.getCode(), 400);
-            assertTrue(ex.getMessage().contains("No templates specified"));
-        }
-
-        final String templateName = "test validate";
-        templates.add(templateName);
-        try {
-            zmsImpl.putDomainTemplateExt(ctx, domainName, templateName,
-                    auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(400, ex.getCode());
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateExtNotFoundTemplate() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        String domainName = "templatelist-invalid";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        final String templateName = "InvalidTemplate";
-        DomainTemplate templateList = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templates.add(templateName);
-        templateList.setTemplateNames(templates);
-        try {
-            zmsImpl.putDomainTemplateExt(ctx, domainName, templateName,
-                    auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(404, ex.getCode());
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateExtEmptyTemplateList() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        String domainName = "templatelist-empty";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        DomainTemplate templateList = new DomainTemplate();
-        try {
-            zmsImpl.putDomainTemplateExt(ctx, domainName, "unknown",
-                    auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(400, ex.getCode());
-        }
-
-        templateList.setTemplateNames(Collections.emptyList());
-        try {
-            zmsImpl.putDomainTemplateExt(ctx, domainName, "unknown",
-                    auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(400, ex.getCode());
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateExtMultipleTemplate() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-        String domainName = "templatelist-invalid";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        final String templateName = "vipng";
-        DomainTemplate templateList = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templates.add(templateName);
-        templates.add("pes");
-        templateList.setTemplateNames(templates);
-        try {
-            zmsImpl.putDomainTemplateExt(ctx, domainName, templateName,
-                    auditRef, templateList);
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(400, ex.getCode());
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutDomainTemplateExtSingleTemplate() {
-
-        String domainName = "templatelist-single";
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        SubDomain domSysNetwork = zmsTestInitializer.createSubDomainObject("network", "sys", "Test Domain",
-                "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postSubDomain(ctx, "sys", auditRef, domSysNetwork);
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        final String templateName = "vipng";
-        DomainTemplate domTemplate = new DomainTemplate();
-        List<String> templates = new ArrayList<>();
-        templates.add(templateName);
-        domTemplate.setTemplateNames(templates);
-
-        zmsImpl.putDomainTemplateExt(ctx, domainName, templateName,
-                auditRef, domTemplate);
-
-        // verify that our role collection includes the roles defined in template
-
-        List<String> names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(3, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-
-        Role role = zmsImpl.dbService.getRole(domainName, "vip_admin", false, false, false);
-        assertEquals(domainName + ":role.vip_admin", role.getName());
-        assertNull(role.getTrust());
-        assertTrue(role.getRoleMembers().isEmpty());
-
-        role = zmsImpl.dbService.getRole(domainName, "sys_network_super_vip_admin", false, false, false);
-        assertEquals(domainName + ":role.sys_network_super_vip_admin", role.getName());
-        assertEquals("sys.network", role.getTrust());
-
-        // verify that our policy collections includes the policies defined in the template
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(3, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-
-        Policy policy = zmsImpl.dbService.getPolicy(domainName, "vip_admin", null);
-        assertEquals(domainName + ":policy.vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        Assertion assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(domainName + ":role.vip_admin", assertion.getRole());
-        assertEquals(domainName + ":vip*", assertion.getResource());
-
-        policy = zmsImpl.dbService.getPolicy(domainName, "sys_network_super_vip_admin", null);
-        assertEquals(domainName + ":policy.sys_network_super_vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(domainName + ":role.sys_network_super_vip_admin", assertion.getRole());
-        assertEquals(domainName + ":vip*", assertion.getResource());
-
-        // delete an applied service template
-        //
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName, auditRef);
-
-        // verify that our role collection does NOT include the roles defined in template
-
-        names = zmsImpl.dbService.listRoles(domainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        names = zmsImpl.dbService.listPolicies(domainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        zmsImpl.deleteSubDomain(ctx, "sys", "network", auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testGetDomainTemplateListInvalid() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        try {
-            zmsImpl.getDomainTemplateList(ctx, "invalid_domain name");
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(400, ex.getCode());
-        }
-
-        try {
-            zmsImpl.getDomainTemplateList(ctx, "not_found_domain_name");
-            fail();
-        } catch (ResourceException ex) {
-            assertEquals(404, ex.getCode());
-        }
-    }
-
-    @Test
-    public void testGetDomainTemplateList() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        String domainName = "domaintemplatelist-valid";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        // initially no templates
-
-        DomainTemplateList domaintemplateList = zmsImpl.getDomainTemplateList(ctx,
-                domainName);
-        List<String> templates = domaintemplateList.getTemplateNames();
-        assertEquals(0, templates.size());
-
-        // add a single template
-
-        DomainTemplate domTemplate = new DomainTemplate();
-        templates = new ArrayList<>();
-        templates.add("user_provisioning");
-        domTemplate.setTemplateNames(templates);
-
-        zmsImpl.putDomainTemplate(ctx, domainName, auditRef,
-                domTemplate);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, domainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(1, templates.size());
-        assertTrue(templates.contains("user_provisioning"));
-
-        // add 2 templates
-
-        domTemplate = new DomainTemplate();
-        templates = new ArrayList<>();
-        templates.add("user_provisioning");
-        templates.add("platforms");
-        domTemplate.setTemplateNames(templates);
-
-        zmsImpl.putDomainTemplate(ctx, domainName, auditRef,
-                domTemplate);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, domainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(2, templates.size());
-        assertTrue(templates.contains("user_provisioning"));
-        assertTrue(templates.contains("platforms"));
-
-        // add the same set of templates again and no change in results
-        domTemplate = new DomainTemplate();
-        domTemplate.setTemplateNames(templates);
-        zmsImpl.putDomainTemplate(ctx, domainName, auditRef,
-                domTemplate);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, domainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(2, templates.size());
-        assertTrue(templates.contains("user_provisioning"));
-        assertTrue(templates.contains("platforms"));
-
-        // delete an applied service template
-        //
-        String templateName = "user_provisioning";
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName,
-                auditRef);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, domainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(1, templates.size());
-        assertTrue(templates.contains("platforms"));
-
-        // delete last applied service template
-        //
-        templateName = "platforms";
-        zmsImpl.deleteDomainTemplate(ctx, domainName, templateName,
-                auditRef);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, domainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertTrue(templates.isEmpty());
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPostSubDomainWithTemplates() {
-
-        String domainName = "postsubdomain-withtemplate";
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        SubDomain domSysNetwork = zmsTestInitializer.createSubDomainObject("network", "sys", "Test Domain",
-                "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postSubDomain(ctx, "sys", auditRef, domSysNetwork);
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        SubDomain dom2 = zmsTestInitializer.createSubDomainObject("sub", domainName,
-                "Test Domain2", "testOrg", zmsTestInitializer.getAdminUser());
-        DomainTemplateList templateList = new DomainTemplateList();
-        List<String> templates = new ArrayList<>();
-        templates.add("vipng");
-        templates.add("platforms");
-        templates.add("user_provisioning");
-        templateList.setTemplateNames(templates);
-        dom2.setTemplates(templateList);
-
-        Domain resDom1 = zmsImpl.postSubDomain(ctx, domainName,
-                auditRef, dom2);
-        assertNotNull(resDom1);
-
-        String subDomainName = domainName + ".sub";
-
-        // verify that our role collection includes the roles defined in template
-
-        List<String> names = zmsImpl.dbService.listRoles(subDomainName);
-        assertEquals(7, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-        assertTrue(names.contains("platforms_deployer"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        Role role = zmsImpl.dbService.getRole(subDomainName, "openstack_readers", false, false, false);
-        assertEquals(subDomainName + ":role.openstack_readers", role.getName());
-        assertNull(role.getTrust());
-        assertEquals(2, role.getRoleMembers().size());
-
-        List<String> checkList = new ArrayList<>();
-        checkList.add("sys.builder");
-        checkList.add("sys.openstack");
-        zmsTestInitializer.checkRoleMember(checkList, role.getRoleMembers());
-
-        role = zmsImpl.dbService.getRole(subDomainName, "sys_network_super_vip_admin", false, false, false);
-        assertEquals(subDomainName + ":role.sys_network_super_vip_admin", role.getName());
-        assertEquals("sys.network", role.getTrust());
-
-        // verify that our policy collections includes the policies defined in the template
-
-        names = zmsImpl.dbService.listPolicies(subDomainName);
-        assertEquals(7, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("vip_admin"));
-        assertTrue(names.contains("sys_network_super_vip_admin"));
-        assertTrue(names.contains("platforms_deploy"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        Policy policy = zmsImpl.dbService.getPolicy(subDomainName, "vip_admin", null);
-        assertEquals(subDomainName + ":policy.vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        Assertion assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(subDomainName + ":role.vip_admin", assertion.getRole());
-        assertEquals(subDomainName + ":vip*", assertion.getResource());
-
-        policy = zmsImpl.dbService.getPolicy(subDomainName, "sys_network_super_vip_admin", null);
-        assertEquals(subDomainName + ":policy.sys_network_super_vip_admin", policy.getName());
-        assertEquals(1, policy.getAssertions().size());
-        assertion = policy.getAssertions().get(0);
-        assertEquals("*", assertion.getAction());
-        assertEquals(subDomainName + ":role.sys_network_super_vip_admin", assertion.getRole());
-        assertEquals(subDomainName + ":vip*", assertion.getResource());
-
-        // verify the saved domain list
-
-        DomainTemplateList domaintemplateList = zmsImpl.getDomainTemplateList(ctx, subDomainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(3, templates.size());
-        assertTrue(templates.contains("vipng"));
-        assertTrue(templates.contains("platforms"));
-        assertTrue(templates.contains("user_provisioning"));
-
-        // delete an applied service template
-        //
-        String templateName = "vipng";
-        zmsImpl.deleteDomainTemplate(ctx, subDomainName, templateName, auditRef);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, subDomainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(2, templates.size());
-        assertTrue(templates.contains("platforms"));
-        assertTrue(templates.contains("user_provisioning"));
-
-        names = zmsImpl.dbService.listRoles(subDomainName);
-        assertEquals(5, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("platforms_deployer"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        names = zmsImpl.dbService.listPolicies(subDomainName);
-        assertEquals(5, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("platforms_deploy"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        // delete an applied service template
-        //
-        templateName = "platforms";
-        zmsImpl.deleteDomainTemplate(ctx, subDomainName, templateName, auditRef);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, subDomainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertEquals(1, templates.size());
-        assertTrue(templates.contains("user_provisioning"));
-
-        names = zmsImpl.dbService.listRoles(subDomainName);
-        assertEquals(4, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        names = zmsImpl.dbService.listPolicies(subDomainName);
-        assertEquals(4, names.size());
-        assertTrue(names.contains("admin"));
-        assertTrue(names.contains("user"));
-        assertTrue(names.contains("superuser"));
-        assertTrue(names.contains("openstack_readers"));
-
-        // delete last applied service template
-        //
-        templateName = "user_provisioning";
-        zmsImpl.deleteDomainTemplate(ctx, subDomainName, templateName, auditRef);
-
-        domaintemplateList = zmsImpl.getDomainTemplateList(ctx, subDomainName);
-        templates = domaintemplateList.getTemplateNames();
-        assertTrue(templates.isEmpty());
-
-        names = zmsImpl.dbService.listRoles(subDomainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        names = zmsImpl.dbService.listPolicies(subDomainName);
-        assertEquals(1, names.size());
-        assertTrue(names.contains("admin"));
-
-        zmsImpl.deleteSubDomain(ctx, "sys", "network", auditRef);
-        zmsImpl.deleteSubDomain(ctx, domainName, "sub", auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
     public void testPutPolicyNoLoopbackNoSuchDomainError() {
         final String auditRef = zmsTestInitializer.getAuditRef();
         HttpServletRequest servletRequest = Mockito.mock(HttpServletRequest.class);
@@ -18520,6 +17732,7 @@ public class ZMSImplTest {
         // TODO this is only a skeleton, we need to add tests once the validateAssertionConditions func implemented
     }
 
+    @Test
     public void testSetupRoleListWithMembers() {
 
         ZMSImpl zmsImpl = zmsTestInitializer.getZms();
@@ -26185,7 +25398,7 @@ public class ZMSImplTest {
 
         Response response = zmsImpl.getJWSDomain(ctx, domainName, null, null);
         JWSDomain jwsDomain = (JWSDomain) response.getEntity();
-        DomainData domainData = getDomainData(jwsDomain);
+        DomainData domainData = zmsTestInitializer.getDomainData(jwsDomain);
 
         assertNotNull(domainData);
         assertEquals(domainData.getName(), "jws-domain");
@@ -26207,7 +25420,7 @@ public class ZMSImplTest {
         Timestamp tstamp = Timestamp.fromMillis(System.currentTimeMillis() - 3600);
         response = zmsImpl.getJWSDomain(ctx, domainName, false, tstamp.toString());
         jwsDomain = (JWSDomain) response.getEntity();
-        domainData = getDomainData(jwsDomain);
+        domainData = zmsTestInitializer.getDomainData(jwsDomain);
 
         assertNotNull(domainData);
         assertEquals(domainData.getName(), "jws-domain");
@@ -26217,7 +25430,7 @@ public class ZMSImplTest {
 
         response = zmsImpl.getJWSDomain(ctx, domainName, null, "unknown-date");
         jwsDomain = (JWSDomain) response.getEntity();
-        domainData = getDomainData(jwsDomain);
+        domainData = zmsTestInitializer.getDomainData(jwsDomain);
 
         assertNotNull(domainData);
         assertEquals(domainData.getName(), "jws-domain");
@@ -26564,10 +25777,17 @@ public class ZMSImplTest {
         RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
 
         DomainTemplateDetailsList serverTemplateDetailsList = zmsImpl.getServerTemplateDetailsList(ctx);
-        assertEquals(serverTemplateDetailsList.getMetaData().size(), 10);
-        assertEquals(serverTemplateDetailsList.getMetaData().get(6).getDescription(), "Vipng template");
-        assertEquals(serverTemplateDetailsList.getMetaData().get(6).getLatestVersion().intValue(), 10);
-        assertEquals(serverTemplateDetailsList.getMetaData().get(6).getTemplateName(), "vipng");
+        assertEquals(serverTemplateDetailsList.getMetaData().size(), 11);
+        TemplateMetaData vipTemplateMetaData = null;
+        for (TemplateMetaData templateMetaData : serverTemplateDetailsList.getMetaData()) {
+            if (templateMetaData.getTemplateName().equals("vipng")) {
+                vipTemplateMetaData = templateMetaData;
+                break;
+            }
+        }
+        assertNotNull(vipTemplateMetaData);
+        assertEquals(vipTemplateMetaData.getDescription(), "Vipng template");
+        assertEquals(vipTemplateMetaData.getLatestVersion().intValue(), 10);
     }
 
     @Test
@@ -31574,582 +30794,6 @@ public class ZMSImplTest {
     }
 
     @Test
-    public void testQueryPutRoleWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "sys.auth";
-
-        // put role with multiple tags
-        final String roleWithTags = "roleWithTags";
-        final String tagKey = "tag-key";
-        List<String> tagValues = Arrays.asList("val1", "val2");
-        Role role = zmsTestInitializer.createRoleObject(domainName, roleWithTags, null);
-        role.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(tagValues)));
-        zmsImpl.putRole(ctx, domainName, roleWithTags, auditRef, false, role);
-
-        // put role with single tags
-        final String roleSingleTag = "roleSingleTag";
-        List<String> singleTagValue = Collections.singletonList("val1");
-        role = zmsTestInitializer.createRoleObject(domainName, roleSingleTag, null);
-        role.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(singleTagValue)));
-        zmsImpl.putRole(ctx, domainName, roleSingleTag, auditRef, false, role);
-
-        //put role without tags
-        final String noTagsRole = "noTagsRole";
-        role = zmsTestInitializer.createRoleObject(domainName, noTagsRole, null);
-        zmsImpl.putRole(ctx, domainName, noTagsRole, auditRef, false, role);
-
-        // get roles without tags query - both tags should be presented
-        Roles roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, null, null);
-        hasRoleWithTags(roleList, roleWithTags, tagKey, tagValues, 2);
-        hasRoleWithTags(roleList, roleSingleTag, tagKey, singleTagValue, 1);
-        hasRoleWithTags(roleList, noTagsRole, null, null, 0);
-
-        // get roles with exact tag value
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, "val1");
-        hasRoleWithTags(roleList, roleWithTags, tagKey, tagValues, 2);
-        hasRoleWithTags(roleList, roleSingleTag, tagKey, singleTagValue, 1);
-        // ensure there are no more roles
-        assertEquals(roleList.getList().size(), 2);
-
-        // get roles with exact tag value
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, "val2");
-        hasRoleWithTags(roleList, roleWithTags, tagKey, tagValues, 2);
-        // ensure there are no more roles
-        assertEquals(roleList.getList().size(), 1);
-
-        // get roles with only tag key
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, null);
-        hasRoleWithTags(roleList, roleWithTags, tagKey, tagValues, 2);
-        hasRoleWithTags(roleList, roleSingleTag, tagKey, singleTagValue, 1);
-        // ensure there are no more roles
-        assertEquals(roleList.getList().size(), 2);
-    }
-
-    @Test
-    public void testRoleTagsLimit() {
-
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        // define limit of 3 role tags
-        System.setProperty(ZMSConsts.ZMS_PROP_QUOTA_ROLE_TAG, "3");
-        ZMSImpl zmsTest = zmsTestInitializer.zmsInit();
-
-        final String domainName = "sys.auth";
-        final String roleName = "roleWithTagLimit";
-        final String tagKey = "tag-key";
-
-        //insert role with 4 tags
-        List<String> tagValues = Arrays.asList("val1", "val2", "val3", "val4");
-        Role role = zmsTestInitializer.createRoleObject(domainName, roleName, null);
-        role.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(tagValues)));
-        try {
-            zmsTest.putRole(ctx, domainName, roleName, auditRef, false, role);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.BAD_REQUEST);
-            assertTrue(ex.getMessage().contains("role tag quota exceeded - limit: 3, current tags count: 0, new tags count: 4"));
-        }
-
-        try {
-            // role should not be created if fails to process tags..
-            zmsTest.getRole(ctx, domainName, roleName, false, false, false);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.NOT_FOUND);
-        }
-
-        System.clearProperty(ZMSConsts.ZMS_PROP_QUOTA_ROLE_TAG);
-    }
-
-    @Test
-    public void testQueryUpdateRoleWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "sys.auth";
-        final String tagKey = "tag-key-update";
-
-        //put role without tags
-        final String noTagsRole = "noTagsRole";
-        Role role = zmsTestInitializer.createRoleObject(domainName, noTagsRole, null);
-        zmsImpl.putRole(ctx, domainName, noTagsRole, auditRef, false, role);
-
-        // assert there are no tags
-        Roles roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, null, null);
-        hasRoleWithTags(roleList, noTagsRole, null, null, 0);
-
-        // update tag list
-        List<String> tagValues = Arrays.asList("val1", "val2", "val3");
-        role.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(tagValues)));
-        zmsImpl.putRole(ctx, domainName, noTagsRole, auditRef, false, role);
-
-        // 2 tags should be presented
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, null, null);
-        hasRoleWithTags(roleList, noTagsRole, tagKey, tagValues, 3);
-
-        // get roles with exact tag value
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.FALSE, tagKey, "val1");
-        hasRoleWithTags(roleList, noTagsRole, tagKey, tagValues, 3);
-        assertEquals(roleList.getList().size(), 1);
-
-        // get roles with only tag key
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, null);
-        hasRoleWithTags(roleList, noTagsRole, tagKey, tagValues, 3);
-        assertEquals(roleList.getList().size(), 1);
-
-        // now create a different tags Map, part is from tagValues
-        Map<String, TagValueList> tagsMap = new HashMap<>();
-        List<String> modifiedTagValues = Arrays.asList("val1", "new-val");
-        String newTagKey = "newTagKey";
-        List<String> newTagValues = Arrays.asList("val4", "val5", "val6");
-        tagsMap.put(tagKey, new TagValueList().setList(modifiedTagValues));
-        tagsMap.put(newTagKey, new TagValueList().setList(newTagValues));
-        role.setTags(tagsMap);
-        zmsImpl.putRole(ctx, domainName, noTagsRole, auditRef, false, role);
-
-        // 1 tags should be presented
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, null, null);
-        hasRoleWithTags(roleList, noTagsRole, tagKey, modifiedTagValues, 2);
-        hasRoleWithTags(roleList, noTagsRole, newTagKey, newTagValues, 3);
-
-        // get roles with exact tag value
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, "val1");
-        hasRoleWithTags(roleList, noTagsRole, tagKey, modifiedTagValues, 2);
-        assertEquals(roleList.getList().size(), 1);
-
-        // get roles with non-existent tag value
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, "val2");
-        assertEquals(roleList.getList().size(), 0);
-
-        // get roles with new tag key
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, null);
-        hasRoleWithTags(roleList, noTagsRole, newTagKey, newTagValues, 3);
-        assertEquals(roleList.getList().size(), 1);
-    }
-
-    @Test
-    public void testUpdateRoleMetaWithoutTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "update-role-meta-without-tags";
-        final String updateRoleMetaTag = "tag-key-update-role-meta";
-        final List<String> updateRoleMetaTagValues = Collections.singletonList("update-meta-value");
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        // put role without tags
-        final String roleName = "roleTagsUpdateMeta";
-        Role role = zmsTestInitializer.createRoleObject(domainName, roleName, null);
-        zmsImpl.putRole(ctx, domainName, roleName, auditRef, false, role);
-
-        // no tags should be presented
-        Roles roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, updateRoleMetaTag, null);
-        assertTrue(roleList.getList().isEmpty());
-
-        RoleMeta rm = new RoleMeta()
-                .setTags(Collections.singletonMap(updateRoleMetaTag,
-                        new TagValueList().setList(updateRoleMetaTagValues)));
-
-        // update role tags using role meta
-        zmsImpl.putRoleMeta(ctx, domainName, roleName, auditRef, rm);
-
-        // assert that updateRoleMetaTag is in role tags
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, updateRoleMetaTag, null);
-        hasRoleWithTags(roleList, roleName, updateRoleMetaTag, updateRoleMetaTagValues, 1);
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testUpdateRoleMetaWithExistingTag() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "update-role-meta-with-existing-tag";
-        final String tagKey = "tag-key";
-        final String updateRoleMetaTag = "tag-key-update-role-meta-exist-tag";
-        final List<String> updateRoleMetaTagValues = Collections.singletonList("update-meta-value");
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        // put role with tag
-        final String roleName = "roleWithTagUpdateMeta";
-        List<String> singleTagValue = Collections.singletonList("val1");
-        Role role = zmsTestInitializer.createRoleObject(domainName, roleName, null);
-        role.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(singleTagValue)));
-        zmsImpl.putRole(ctx, domainName, roleName, auditRef, false, role);
-
-        // tag tagKey should be presented
-        Roles roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, tagKey, null);
-        hasRoleWithTags(roleList, roleName, tagKey, singleTagValue, 1);
-
-        RoleMeta rm = new RoleMeta()
-                .setTags(Collections.singletonMap(updateRoleMetaTag,
-                        new TagValueList().setList(updateRoleMetaTagValues)));
-
-        // update role tags using role meta
-        zmsImpl.putRoleMeta(ctx, domainName, roleName, auditRef, rm);
-
-        // role should contain only the new tag
-        roleList = zmsImpl.getRoles(ctx, domainName, Boolean.TRUE, updateRoleMetaTag, null);
-        hasRoleWithTags(roleList, roleName, updateRoleMetaTag, updateRoleMetaTagValues, 1);
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    private void hasRoleWithTags(Roles roleList, String roleName, String tagKey, List<String> tagValues, int tagValuesLength) {
-        Role role = getRole(roleList, roleName);
-        Assert.assertNotNull(role);
-        if (tagKey != null) {
-            if (tagValues != null) {
-                Assert.assertEquals(role.getTags().get(tagKey).getList().size(), tagValuesLength);
-                for (String tagValue : tagValues) {
-                    Assert.assertTrue(hasTag(role, tagKey, tagValue));
-                }
-            } else {
-                Assert.assertTrue(hasTag(role, tagKey, null));
-            }
-        }
-    }
-
-    private void hasPolicyWithTags(Policies policies, String policyName, String version, String tagKey, List<String> tagValues, int tagValuesLength) {
-        Policy policy = getPolicy(policies, policyName, version);
-        Assert.assertNotNull(policy);
-        if (tagKey != null) {
-            if (tagValues != null) {
-                Assert.assertEquals(policy.getTags().get(tagKey).getList().size(), tagValuesLength);
-                for (String tagValue : tagValues) {
-                    Assert.assertTrue(verifyPolicyHasTag(policy, tagKey, tagValue));
-                }
-            } else {
-                Assert.assertTrue(verifyPolicyHasTag(policy, tagKey, null));
-            }
-        }
-    }
-
-    private boolean hasTag(Role role, String tagKey, String tagValue) {
-        TagValueList tagValues = role.getTags().get(tagKey);
-        if (tagValue != null) {
-            return tagValues.getList().contains(tagValue);
-        }
-        return !tagValues.getList().isEmpty();
-    }
-
-    private Role getRole(Roles roleList, String roleName) {
-        return roleList.getList().stream()
-                .filter(r -> AthenzUtils.extractRoleName(r.getName()).equalsIgnoreCase(roleName))
-                .findFirst()
-                .get();
-    }
-
-    private boolean verifyPolicyHasTag(Policy policy, String tagKey, String tagValue) {
-        TagValueList tagValues = policy.getTags().get(tagKey);
-        if (tagValue != null) {
-            return tagValues.getList().contains(tagValue);
-        }
-        return !tagValues.getList().isEmpty();
-    }
-
-    private Policy getPolicy(Policies polices, String policyName, String version) {
-        return polices.getList().stream()
-                .filter(r -> AthenzUtils.extractPolicyName(r.getName()).equalsIgnoreCase(policyName) && r.getVersion().equalsIgnoreCase(version))
-                .findFirst()
-                .get();
-    }
-
-    @Test
-    public void testDomainMetaWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "domain-with-tags";
-
-        TopLevelDomain topLevelDomain = zmsTestInitializer.createTopLevelDomainObject(domainName, "Test Domain With Tags", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, topLevelDomain);
-
-        DomainMeta domainMeta = zmsTestInitializer.createDomainMetaObject("Domain Meta for domain tags", "testOrg", true, true, "12345", 1001);
-        domainMeta.setTags(simpleDomainTag());
-        zmsImpl.putDomainMeta(ctx, domainName, auditRef, domainMeta);
-
-        Domain domain = zmsImpl.getDomain(ctx, domainName);
-        assertEquals(domain.getTags(), simpleDomainTag());
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testDomainTagsLimit() throws Exception {
-        // define limit of 3 domain tags
-        System.setProperty(ZMSConsts.ZMS_PROP_QUOTA_DOMAIN_TAG, "3");
-        ZMSImpl zmsTest = zmsTestInitializer.zmsInit();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-        final String domainName = "tld-with-tag-limit";
-
-        TopLevelDomain topLevelDomain = zmsTestInitializer.createTopLevelDomainObject(domainName, "Test Domain With Tag Limit", "testOrg", zmsTestInitializer.getAdminUser());
-        topLevelDomain.setTags(Collections.singletonMap("tag-key", new TagValueList().setList(Arrays.asList("val1", "val2", "val3", "val4"))));
-        try {
-            zmsTest.postTopLevelDomain(ctx, auditRef, topLevelDomain);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.BAD_REQUEST);
-            assertTrue(ex.getMessage().contains("domain tag quota exceeded - limit: 3, current tags count: 0, new tags count: 4"));
-        }
-        try {
-            // domain should not be created if fails to process tags..
-            zmsTest.getDomain(ctx, domainName);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.NOT_FOUND);
-        }
-
-        System.clearProperty(ZMSConsts.ZMS_PROP_QUOTA_DOMAIN_TAG);
-    }
-
-    @Test
-    public void testTopLevelSubDomainWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "tld-with-tags";
-
-        TopLevelDomain topLevelDomain = zmsTestInitializer.createTopLevelDomainObject(domainName, "Test Domain With Tags", "testOrg", zmsTestInitializer.getAdminUser());
-        topLevelDomain.setTags(simpleDomainTag());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, topLevelDomain);
-
-        String subDomName = "subdomain-with-tag";
-        SubDomain subDom = zmsTestInitializer.createSubDomainObject(subDomName, domainName, "subdomain desc", "testOrg", zmsTestInitializer.getAdminUser());
-        subDom.setTags(simpleDomainTag());
-        zmsImpl.postSubDomain(ctx, domainName, auditRef, subDom);
-
-        Domain domain = zmsImpl.getDomain(ctx, domainName);
-        assertEquals(domain.getTags(), simpleDomainTag());
-
-        Domain subDomainObj = zmsImpl.getDomain(ctx, domainName + "." + subDomName);
-        assertEquals(subDomainObj.getTags(), simpleDomainTag());
-
-        zmsImpl.deleteSubDomain(ctx, domainName, subDomName, auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testUserLevelDomainWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        String domainName = "dguttman-tags";
-
-        UserDomain userDomain = zmsTestInitializer.createUserDomainObject(domainName, "Test Domain1", "testOrg");
-        userDomain.setTags(simpleDomainTag());
-        zmsImpl.postUserDomain(ctx, domainName, auditRef, userDomain);
-
-        Domain domain = zmsImpl.getDomain(ctx, "user." + domainName);
-        assertEquals(domain.getTags(), simpleDomainTag());
-
-        zmsImpl.deleteUserDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testGetSignedDomainWithTags() throws JsonProcessingException, ParseException, JOSEException {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "jws-domain-tags";
-
-        // create multiple top level domains
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName, "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        dom1.setTags(simpleDomainTag());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Response response = zmsImpl.getJWSDomain(ctx, domainName, null, null);
-        JWSDomain jwsDomain = (JWSDomain) response.getEntity();
-        DomainData domainData = getDomainData(jwsDomain);
-        assertNotNull(domainData);
-        assertEquals(domainData.getName(), domainName);
-
-        Map<String, String> header = jwsDomain.getHeader();
-        assertEquals(header.get("kid"), "0");
-
-        assertEquals(domainData.getTags(), simpleDomainTag());
-
-        Authority principalAuthority = new com.yahoo.athenz.common.server.debug.DebugPrincipalAuthority();
-        Principal sysPrincipal = principalAuthority.authenticate("v=U1;d=sys;n=zts;s=signature",
-                "10.11.12.13", "GET", null);
-        ResourceContext rsrcCtx = zmsTestInitializer.createResourceContext(sysPrincipal);
-
-        response = zmsImpl.getSignedDomains(rsrcCtx, domainName, null, null, null, false, null);
-        SignedDomains sdoms = (SignedDomains) response.getEntity();
-        assertNotNull(sdoms);
-
-        Map<String, TagValueList> signedDomainTags = sdoms.getDomains().stream()
-                .filter(dom -> dom.getDomain().getName().equals(domainName))
-                .map(dom -> dom.getDomain().getTags())
-                .findFirst().get();
-
-        assertEquals(signedDomainTags, simpleDomainTag());
-
-        // test with meta only
-        response = zmsImpl.getSignedDomains(rsrcCtx, domainName, "true", "all", null, false, null);
-        sdoms = (SignedDomains) response.getEntity();
-        assertNotNull(sdoms);
-
-        signedDomainTags = sdoms.getDomains().stream()
-                .filter(dom -> dom.getDomain().getName().equals(domainName))
-                .map(dom -> dom.getDomain().getTags())
-                .findFirst().get();
-
-        assertEquals(signedDomainTags, simpleDomainTag());
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    private DomainData getDomainData(JWSDomain jwsDomain) throws ParseException, JOSEException, JsonProcessingException {
-        assertNotNull(jwsDomain);
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-
-        JWSObject jwsObject = new JWSObject(Base64URL.from(jwsDomain.getProtectedHeader()),
-                Base64URL.from(jwsDomain.getPayload()), Base64URL.from(jwsDomain.getSignature()));
-
-        JWSVerifier verifier = new RSASSAVerifier((RSAPublicKey) Crypto.extractPublicKey(zmsImpl.privateKey.getKey()));
-        assertTrue(jwsObject.verify(verifier));
-
-        return zmsImpl.jsonMapper.readValue(jwsObject.getPayload().toString(), DomainData.class);
-    }
-
-    private Map<String, TagValueList> simpleDomainTag() {
-        return Collections.singletonMap("tag-key", new TagValueList().setList(Arrays.asList("val1", "val2")));
-    }
-
-    @Test
-    public void testGetDomainListUsingTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        // first domain - no tags
-        String domainNoTags = "tld-no-tags";
-        TopLevelDomain tldNoTags = zmsTestInitializer.createTopLevelDomainObject(domainNoTags, "Test Domain Without Tags", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, tldNoTags);
-        Domain domainObjNoTags = zmsImpl.getDomain(ctx, domainNoTags);
-        assertNull(domainObjNoTags.getTags());
-
-        // first domain - 1 tag
-        String domainName1 = "tld-tag-1";
-        TopLevelDomain topLevelDomain = zmsTestInitializer.createTopLevelDomainObject(domainName1, "Test Domain With Tags", "testOrg", zmsTestInitializer.getAdminUser());
-        topLevelDomain.setTags(simpleDomainTag());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, topLevelDomain);
-        Domain domain1 = zmsImpl.getDomain(ctx, domainName1);
-        assertEquals(domain1.getTags(), simpleDomainTag());
-
-        // second domain - 2 tags
-        Map<String, TagValueList> twoTags = new HashMap<>();
-        twoTags.put("tag-key", new TagValueList().setList(Arrays.asList("tld2-val1", "tld2-val2")));
-        twoTags.put("tag-key-2", new TagValueList().setList(Arrays.asList("tld2-val3", "tld2-val4")));
-        String domainName2 = "tld-tag-2";
-        TopLevelDomain topLevelDomain2 = zmsTestInitializer.createTopLevelDomainObject(domainName2, "Test Domain With Tags", "testOrg", zmsTestInitializer.getAdminUser());
-        topLevelDomain2.setTags(twoTags);
-        zmsImpl.postTopLevelDomain(ctx, auditRef, topLevelDomain2);
-        Domain domain2 = zmsImpl.getDomain(ctx, domainName1);
-        assertEquals(domain2.getTags(), simpleDomainTag());
-
-        // domain-list no tags - all domains should be presented
-        DomainList dl = zmsImpl.getDomainList(ctx, null, null, null, null,
-                null, null, null, null, null, null, null, null, null, null, null);
-        assertTrue(dl.getNames().containsAll(Arrays.asList(domainNoTags, domainName1, domainName2)));
-
-        // domain-list with only tag-key, should include both domains
-        dl = zmsImpl.getDomainList(ctx, null, null, null, null,
-                null, null, null, null, null, null, "tag-key", null, null, null, null);
-
-        assertEquals(dl.getNames().size(), 2);
-        assertTrue(dl.getNames().containsAll(Arrays.asList(domainName1, domainName2)));
-
-        // domain-list with tag-key AND tag-value, should include only domainName1
-        dl = zmsImpl.getDomainList(ctx, null, null, null, null,
-                null, null, null, null, null, null, "tag-key", "val1", null, null, null);
-
-        assertEquals(dl.getNames().size(), 1);
-        assertTrue(dl.getNames().contains(domainName1));
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainNoTags, auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName1, auditRef);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName2, auditRef);
-    }
-
-    @Test
-    public void testUpdateDomainTag() {
-        final String domainName = "domain-with-tags";
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        TopLevelDomain topLevelDomain = zmsTestInitializer.createTopLevelDomainObject(domainName, "Test Domain With Tags", "testOrg", zmsTestInitializer.getAdminUser());
-        topLevelDomain.setTags(simpleDomainTag());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, topLevelDomain);
-
-        // domain should contain the tag
-        Domain domain = zmsImpl.getDomain(ctx, domainName);
-        assertEquals(domain.getTags(), simpleDomainTag());
-
-        // update domain meta with the same tag..
-        DomainMeta domainMeta = zmsTestInitializer.createDomainMetaObject("Domain Meta for domain tags", "testOrg", true, true, "12345", 1001);
-        domainMeta.setTags(simpleDomainTag());
-        zmsImpl.putDomainMeta(ctx, domainName, auditRef, domainMeta);
-
-        // should be the same tag result..
-        domain = zmsImpl.getDomain(ctx, domainName);
-        assertEquals(domain.getTags(), simpleDomainTag());
-
-        // update domain meta with the same tag, key, but different values..
-        domainMeta = zmsTestInitializer.createDomainMetaObject("Domain Meta for domain tags", "testOrg", true, true, "12345", 1001);
-        Map<String, TagValueList> newTags = Collections.singletonMap("tag-key", new TagValueList().setList(Arrays.asList("val2", "val3")));
-        domainMeta.setTags(newTags);
-        zmsImpl.putDomainMeta(ctx, domainName, auditRef, domainMeta);
-
-        // should be the newTags
-        domain = zmsImpl.getDomain(ctx, domainName);
-        assertEquals(domain.getTags(), newTags);
-
-        // update domain meta with the different tags
-        domainMeta = zmsTestInitializer.createDomainMetaObject("Domain Meta for domain tags", "testOrg", true, true, "12345", 1001);
-        Map<String, TagValueList> newTags2 = Collections.singletonMap("tag-key-2", new TagValueList().setList(Arrays.asList("new-val1", "new-val2")));
-        domainMeta.setTags(newTags2);
-        zmsImpl.putDomainMeta(ctx, domainName, auditRef, domainMeta);
-
-        // should be the newTags2
-        domain = zmsImpl.getDomain(ctx, domainName);
-        assertEquals(domain.getTags(), newTags2);
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
     public void testGetAthenzDomainWithEntities() {
 
         ZMSImpl zmsImpl = zmsTestInitializer.getZms();
@@ -34311,469 +32955,5 @@ public class ZMSImplTest {
         assertFalse(zmsImpl.validateGcpProjectDetails(null, "1234"));
         assertFalse(zmsImpl.validateGcpProjectDetails("1234", ""));
         assertFalse(zmsImpl.validateGcpProjectDetails("1234", null));
-    }
-
-    @Test
-    public void testGetPolicyWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "setup-policy-with-tags";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-        Map<String, TagValueList> policy1Tags = Collections.singletonMap("tag-key", new TagValueList().setList(Arrays.asList("val2", "val3")));
-        Policy policy1 = zmsTestInitializer.createPolicyObject(domainName, "policy1");
-        policy1.setTags(policy1Tags);
-        zmsImpl.putPolicy(ctx, domainName, "policy1", auditRef, false, policy1);
-
-        Policy policy1version2 = zmsTestInitializer.createPolicyObject(domainName, "policy1");
-        Map<String, TagValueList> policy2Tags = Collections.singletonMap("tag-key2", new TagValueList().setList(Arrays.asList("val5", "val6")));
-        policy1version2.setVersion("2");
-        policy1version2.setActive(false);
-        policy1version2.setTags(policy2Tags);
-        zmsImpl.putPolicy(ctx, domainName, "policy1", auditRef, false, policy1version2);
-
-        Policy returnedPolicy = zmsImpl.getPolicy(ctx, domainName, "policy1");
-        verifyPolicyHasTag(returnedPolicy, "tag-key", "val2");
-
-        zmsImpl.setActivePolicyVersion(ctx,domainName, "policy1",new PolicyOptions().setFromVersion("0").setVersion("2"), "test");
-
-        Policy returnedPolicy2 = zmsImpl.getPolicy(ctx, domainName, "policy1");
-        verifyPolicyHasTag(returnedPolicy2, "tag-key2", "val6");
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testSetupPolicyListWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "setup-policy-with-tags";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Test Domain1", "testOrg", zmsTestInitializer.getAdminUser());
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-        Map<String, TagValueList> policy1Tags = Collections.singletonMap("tag-key", new TagValueList().setList(Arrays.asList("val2", "val3")));
-        Policy policy1 = zmsTestInitializer.createPolicyObject(domainName, "policy1");
-        policy1.setTags(policy1Tags);
-        zmsImpl.putPolicy(ctx, domainName, "policy1", auditRef, false, policy1);
-
-        Policy policy1version2 = zmsTestInitializer.createPolicyObject(domainName, "policy1");
-        policy1version2.setVersion("2");
-        policy1version2.setActive(false);
-        policy1version2.setTags(policy1Tags);
-        zmsImpl.putPolicy(ctx, domainName, "policy1", auditRef, false, policy1version2);
-
-        Map<String, TagValueList> policy2Tags = Collections.singletonMap("tag-key", new TagValueList().setList(Arrays.asList("val2")));
-        Policy policy2 = zmsTestInitializer.createPolicyObject(domainName, "policy2");
-        policy2.setTags(policy2Tags);
-        zmsImpl.putPolicy(ctx, domainName, "policy2", auditRef, false, policy2);
-
-        AthenzDomain domain = zmsImpl.getAthenzDomain(domainName, false);
-        List<Policy> policies = zmsImpl.setupPolicyList(domain, Boolean.FALSE, Boolean.FALSE, "tag-key", "val3");
-        assertEquals(1, policies.size()); // need to account for admin policy
-
-        assertEquals(policies.get(0).getName(), "setup-policy-with-tags:policy.policy1");
-        assertEquals(policies.get(0).getVersion(), "0");
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-
-    @Test
-    public void testQueryPutPolicyWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "sys.auth";
-        final String defaultPolicyVersion = "0";
-        // put policy with multiple tags
-        final String policyWithTags = "pwt_policyWithTags";
-        final String tagKey = "tag-key";
-        List<String> multipleTagValues = Arrays.asList("val1", "val2");
-        addRoleNeededForTest(domainName, "Role1");
-        Policy policy = zmsTestInitializer.createPolicyObject(domainName, policyWithTags);
-        policy.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(multipleTagValues)));
-        zmsImpl.putPolicy(ctx, domainName, policyWithTags, auditRef, false, policy);
-
-        // put policy with single tags
-        final String policySingleTag = "pwt_policySingleTag";
-        List<String> singleTagValue = Collections.singletonList("val1");
-        policy = zmsTestInitializer.createPolicyObject(domainName, policySingleTag);
-        policy.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(singleTagValue)));
-        zmsImpl.putPolicy(ctx, domainName, policySingleTag, auditRef, false, policy);
-
-        // put policy version 2 with different tags
-        List<String> version2TagValue = Arrays.asList("val2","val3");
-        String policyVersion = "2";
-        policy = zmsTestInitializer.createPolicyObject(domainName, policyWithTags, policyVersion, Boolean.FALSE);
-        policy.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(version2TagValue)));
-        zmsImpl.putPolicy(ctx, domainName, policyWithTags, auditRef, false, policy);
-
-        //put policy without tags
-        final String noTagsPolicy = "pwt_noTagsPolicy";
-        policy = zmsTestInitializer.createPolicyObject(domainName, noTagsPolicy);
-        zmsImpl.putPolicy(ctx, domainName, noTagsPolicy, auditRef, false, policy);
-
-        // get policies without tags query - all 4 policies should be  presented and the admin policy
-        Policies policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasPolicyWithTags(policyList, policyWithTags, defaultPolicyVersion, tagKey, multipleTagValues, 2);
-        hasPolicyWithTags(policyList, policyWithTags, policyVersion, tagKey, version2TagValue, 2);
-        hasPolicyWithTags(policyList, policySingleTag, defaultPolicyVersion, tagKey, singleTagValue, 1);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, null, null, 0);
-        List<Policy> TestPolicies = policyList.getList().stream().filter(policy1 -> AthenzUtils.extractPolicyName(policy1.getName()).startsWith("pwt_")).collect(Collectors.toList());
-        assertEquals(TestPolicies.size(), 4);
-
-        // get policies with exact tag value
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val1");
-        hasPolicyWithTags(policyList, policyWithTags, defaultPolicyVersion, tagKey, multipleTagValues, 2);
-        hasPolicyWithTags(policyList, policySingleTag, defaultPolicyVersion, tagKey, singleTagValue, 1);
-        // ensure there are no more policies
-        assertEquals(policyList.getList().size(), 2);
-
-        // get policies with exact tag value
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val2");
-        hasPolicyWithTags(policyList, policyWithTags, defaultPolicyVersion, tagKey, multipleTagValues, 2);
-        hasPolicyWithTags(policyList, policyWithTags, policyVersion, tagKey, version2TagValue, 2);
-        // ensure there are no more policies
-        assertEquals(policyList.getList().size(), 2);
-
-        // get policies with only tag key without the unactive policy version
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.FALSE, tagKey, null);
-        hasPolicyWithTags(policyList, policyWithTags, defaultPolicyVersion, tagKey, multipleTagValues, 2);
-        hasPolicyWithTags(policyList, policySingleTag, defaultPolicyVersion, tagKey, singleTagValue, 1);
-        // ensure there are no more policies
-        assertEquals(policyList.getList().size(), 2);
-    }
-
-    @Test
-    public void testPolicyTagsLimit() {
-
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        // define limit of 3 policy tags
-        System.setProperty(ZMSConsts.ZMS_PROP_QUOTA_POLICY_TAG, "3");
-        ZMSImpl zmsTest = zmsTestInitializer.zmsInit();
-
-        final String domainName = "sys.auth";
-        final String policyName = "policyWithTagLimit";
-        final String tagKey = "tag-key";
-
-        //insert policy with 4 tags
-        List<String> tagValues = Arrays.asList("val1", "val2", "val3", "val4");
-        Policy policy = zmsTestInitializer.createPolicyObject(domainName, policyName);
-        policy.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(tagValues)));
-        try {
-            zmsTest.putPolicy(ctx, domainName, policyName, auditRef, false, policy);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.BAD_REQUEST);
-            assertTrue(ex.getMessage().contains("policy tag quota exceeded - limit: 3, current tags count: 0, new tags count: 4"));
-        }
-
-        try {
-            // policy should not be created if fails to process tags...
-            zmsTest.getPolicy(ctx, domainName, policyName);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.NOT_FOUND);
-        }
-
-        System.clearProperty(ZMSConsts.ZMS_PROP_QUOTA_POLICY_TAG);
-    }
-
-    @Test
-    public void testQueryUpdatePolicyWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "sys.auth";
-        final String tagKey = "tag-key-update";
-        final String defaultPolicyVersion = "0";
-        addRoleNeededForTest(domainName, "Role1");
-
-        //put policy without tags
-        final String noTagsPolicy = "upwt_noTagsPolicy";
-        Policy policy = zmsTestInitializer.createPolicyObject(domainName, noTagsPolicy);
-        policy.setVersion("0");
-        zmsImpl.putPolicy(ctx, domainName, noTagsPolicy, auditRef, false, policy);
-        policy = zmsImpl.getPolicy(ctx, domainName, noTagsPolicy);
-
-        // assert there are no tags
-        Policies policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion,null, null, 0);
-
-        // update tag list
-        List<String> tagValues = Arrays.asList("val1", "val2", "val3");
-        policy.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(tagValues)));
-        zmsImpl.putPolicy(ctx, domainName, noTagsPolicy, auditRef, false, policy);
-
-        String policy2Version = "2";
-        Policy policyV2 = zmsTestInitializer.createPolicyObject(domainName, noTagsPolicy, policy2Version, false);
-        List<String> policyV2TagValues = Arrays.asList("v2val1", "val3");
-        policyV2.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(policyV2TagValues)));
-        zmsImpl.putPolicy(ctx, domainName, noTagsPolicy, auditRef, false, policyV2);
-
-        // 3 tags should be presented
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, tagKey, tagValues, 3);
-        hasPolicyWithTags(policyList, noTagsPolicy, policy2Version, tagKey, policyV2TagValues, 2);
-        List<Policy> TestPolicies = policyList.getList().stream().filter(policy1 -> AthenzUtils.extractPolicyName(policy1.getName()).startsWith("upwt_")).collect(Collectors.toList());
-        assertEquals(TestPolicies.size(), 2);
-
-        // get policies with exact tag value without unactive versions.
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.FALSE, Boolean.FALSE, tagKey, "val3");
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, tagKey, tagValues, 3);
-        assertEquals(policyList.getList().size(), 1);
-
-        // get policies with only tag key without unactive vesrions.
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.FALSE, tagKey, null);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, tagKey, tagValues, 3);
-        assertEquals(policyList.getList().size(), 1);
-
-        // now create a different tags Map, part is from tagValues
-        Map<String, TagValueList> tagsMap = new HashMap<>();
-        List<String> modifiedTagValues = Arrays.asList("val1", "new-val");
-        String newTagKey = "newTagKey";
-        List<String> newTagValues = Arrays.asList("val4", "val5", "val6");
-        tagsMap.put(tagKey, new TagValueList().setList(modifiedTagValues));
-        tagsMap.put(newTagKey, new TagValueList().setList(newTagValues));
-        policy.setTags(tagsMap);
-        zmsImpl.putPolicy(ctx, domainName, noTagsPolicy, auditRef, false, policy);
-
-        // only the version 0 policy should be updated
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, tagKey, modifiedTagValues, 2);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, newTagKey, newTagValues, 3);
-        hasPolicyWithTags(policyList, noTagsPolicy, policy2Version, tagKey, policyV2TagValues, 2);
-        TestPolicies = policyList.getList().stream().filter(policy1 -> AthenzUtils.extractPolicyName(policy1.getName()).startsWith("upwt_")).collect(Collectors.toList());
-        assertEquals(TestPolicies.size(), 2);
-
-        // get policies with exact tag value only version 0 should be returned
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val1");
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, tagKey, modifiedTagValues, 2);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, newTagKey, newTagValues, 3);
-        assertEquals(policyList.getList().size(), 1);
-
-        // updated policy version 2 to contain only val1 tag
-        policyV2TagValues = Arrays.asList("val1");
-        policyV2.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(policyV2TagValues)));
-        zmsImpl.putPolicy(ctx, domainName, noTagsPolicy, auditRef, false, policyV2);
-
-        // get policies with exact tag value both of the policies versions should be returned
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val1");
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, tagKey, modifiedTagValues, 2);
-        hasPolicyWithTags(policyList, noTagsPolicy, defaultPolicyVersion, newTagKey, newTagValues, 3);
-        hasPolicyWithTags(policyList, noTagsPolicy, policy2Version, tagKey, policyV2TagValues, 1);
-        assertEquals(policyList.getList().size(), 2);
-
-        // get policies with no existing tag value
-        policyList = zmsImpl.getPolicies(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val10");
-        assertEquals(policyList.getList().size(), 0);
-    }
-
-    private void hasServiceWithTags(ServiceIdentities services, String serviceName, String domain, String tagKey, List<String> tagValues, int tagValuesLength) {
-        ServiceIdentity service = getServiceIdentity(services, serviceName, domain);
-        Assert.assertNotNull(service);
-        if (tagKey != null) {
-            if (tagValues != null) {
-                Assert.assertEquals(service.getTags().get(tagKey).getList().size(), tagValuesLength);
-                for (String tagValue : tagValues) {
-                    Assert.assertTrue(verifyServiceHasTag(service, tagKey, tagValue));
-                }
-            } else {
-                Assert.assertTrue(verifyServiceHasTag(service, tagKey, null));
-            }
-        }
-    }
-
-    private boolean verifyServiceHasTag(ServiceIdentity service, String tagKey, String tagValue) {
-        TagValueList tagValues = service.getTags().get(tagKey);
-        if (tagValue != null) {
-            return tagValues.getList().contains(tagValue);
-        }
-        return (tagValues == null || tagValues.getList().isEmpty());
-    }
-
-    private ServiceIdentity getServiceIdentity(ServiceIdentities services, String serviceName, String domain) {
-        return services.getList().stream()
-                .filter(r -> ZMSUtils.extractServiceName(domain, r.getName()).equalsIgnoreCase(serviceName))
-                .findFirst()
-                .get();
-    }
-
-    @Test
-    public void testQueryPutServiceWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "sys.auth";
-        // put service with multiple tags
-        final String serviceWithTags = "swt-serviceWithTags";
-        final String tagKey = "tag-key";
-        List<String> multipleTagValues = Arrays.asList("val1", "val2");
-        ServiceIdentity service = zmsTestInitializer.createServiceObject(domainName, serviceWithTags,
-                "http://localhost", "/usr/bin/java", "root", "users", "host1");
-        service.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(multipleTagValues)));
-        zmsImpl.putServiceIdentity(ctx, domainName, serviceWithTags, auditRef, false, service);
-
-        // put service with single tags
-        final String serviceSingleTag = "swt-serviceSingleTag";
-        List<String> singleTagValue = Collections.singletonList("val1");
-        service = zmsTestInitializer.createServiceObject(domainName, serviceSingleTag,
-                "http://localhost", "/usr/bin/java", "root", "users", "host1");
-        service.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(singleTagValue)));
-        zmsImpl.putServiceIdentity(ctx, domainName, serviceSingleTag, auditRef, false, service);
-
-        //put service without tags
-        final String noTagsService = "swt-noTagsService";
-        service = zmsTestInitializer.createServiceObject(domainName, noTagsService,
-                "http://localhost", "/usr/bin/java", "root", "users", "host1");
-        zmsImpl.putServiceIdentity(ctx, domainName, noTagsService, auditRef, false, service);
-
-        // get services without tags query - all 3 services should be  presented
-        ServiceIdentities serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasServiceWithTags(serviceList, serviceWithTags, domainName, tagKey, multipleTagValues, 2);
-        hasServiceWithTags(serviceList, serviceSingleTag, domainName, tagKey, singleTagValue, 1);
-        hasServiceWithTags(serviceList, noTagsService, domainName, null, null, 0);
-        List<ServiceIdentity> TestPolicies = serviceList.getList().stream().filter(service1 -> ZMSUtils.extractServiceName(domainName, service1.getName()).startsWith("swt-")).collect(Collectors.toList());
-        assertEquals(TestPolicies.size(), 3);
-
-        // get policies with exact tag value
-        serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val1");
-        hasServiceWithTags(serviceList, serviceWithTags, domainName, tagKey, multipleTagValues, 2);
-        hasServiceWithTags(serviceList, serviceSingleTag, domainName, tagKey, singleTagValue, 1);
-        // ensure there are no more policies
-        assertEquals(serviceList.getList().size(), 2);
-
-        // get policies with exact tag value
-        serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val2");
-        hasServiceWithTags(serviceList, serviceWithTags, domainName, tagKey, multipleTagValues, 2);
-        // ensure there are no more policies
-        assertEquals(serviceList.getList().size(), 1);
-
-        // get policies with only tag key without the unactive service version
-        serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.FALSE, tagKey, null);
-        hasServiceWithTags(serviceList, serviceWithTags, domainName, tagKey, multipleTagValues, 2);
-        hasServiceWithTags(serviceList, serviceSingleTag, domainName, tagKey, singleTagValue, 1);
-        // ensure there are no more policies
-        assertEquals(serviceList.getList().size(), 2);
-    }
-
-    @Test
-    public void testServiceTagsLimit() {
-
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        // define limit of 3 service tags
-        System.setProperty(ZMSConsts.ZMS_PROP_QUOTA_SERVICE_TAG, "3");
-        ZMSImpl zmsTest = zmsTestInitializer.zmsInit();
-
-        final String domainName = "sys.auth";
-        final String serviceName = "serviceWithTagLimit";
-        final String tagKey = "tag-key";
-
-        //insert service with 4 tags
-        List<String> tagValues = Arrays.asList("val1", "val2", "val3", "val4");
-        ServiceIdentity service = zmsTestInitializer.createServiceObject(domainName, serviceName,
-                "http://localhost", "/usr/bin/java", "root", "users", "host1");
-        service.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(tagValues)));
-        try {
-            zmsTest.putServiceIdentity(ctx, domainName, serviceName, auditRef, false, service);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.BAD_REQUEST);
-            assertTrue(ex.getMessage().contains("service tag quota exceeded - limit: 3, current tags count: 0, new tags count: 4"));
-        }
-
-        try {
-            // service should not be created if fails to process tags...
-            zmsTest.getServiceIdentity(ctx, domainName, serviceName);
-            fail();
-        } catch(ResourceException ex) {
-            assertEquals(ex.getCode(), ResourceException.NOT_FOUND);
-        }
-
-        System.clearProperty(ZMSConsts.ZMS_PROP_QUOTA_SERVICE_TAG);
-    }
-
-
-    @Test
-    public void testQueryUpdateServiceWithTags() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "sys.auth";
-        final String tagKey = "tag-key-update";
-
-        //put service without tags
-        final String noTagsService = "uswt-noTagsService";
-        final String withTagsService = "uswt-withTagsService";
-
-        HashMap<String, TagValueList> multipleTags = new HashMap<>();
-
-
-        ServiceIdentity service = zmsTestInitializer.createServiceObject(domainName, noTagsService,
-                "http://localhost", "/usr/bin/java", "root", "users", "host1");
-        zmsImpl.putServiceIdentity(ctx, domainName, noTagsService, auditRef, false, service);
-
-
-
-        ServiceIdentity service2 = zmsTestInitializer.createServiceObject(domainName, withTagsService,
-                "http://localhost", "/usr/bin/java", "root", "users", "host1");
-        List<String> tagValues = Arrays.asList("val1", "val2");
-        List<String> tag2Values = Arrays.asList("val3", "val4");
-        multipleTags.put(tagKey, new TagValueList().setList(tagValues));
-        multipleTags.put("tagKey2", new TagValueList().setList(tag2Values));
-        service2.setTags(multipleTags);
-        zmsImpl.putServiceIdentity(ctx, domainName, withTagsService, auditRef, false, service2);
-
-        // assert there are no tags
-        ServiceIdentities serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasServiceWithTags(serviceList, noTagsService, domainName,null, null, 0);
-        hasServiceWithTags(serviceList, withTagsService, domainName,tagKey, multipleTags.get(tagKey).getList(), 2);
-        hasServiceWithTags(serviceList, withTagsService, domainName,"tagKey2", multipleTags.get("tagKey2").getList(), 2);
-
-        // update tag list
-        List<String> updatedTagValues = Arrays.asList("val1", "val2", "val3");
-        service.setTags(Collections.singletonMap(tagKey, new TagValueList().setList(updatedTagValues)));
-        zmsImpl.putServiceIdentity(ctx, domainName, noTagsService, auditRef, false, service);
-
-        multipleTags.remove("tagKey2");
-        List<String> newTagValues = Arrays.asList("val3", "val6","val7");
-        multipleTags.put(tagKey, new TagValueList().setList(newTagValues));
-        service2.setTags(multipleTags);
-        zmsImpl.putServiceIdentity(ctx, domainName, withTagsService, auditRef, false, service2);
-
-        // 3 tags should be presented
-        serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.TRUE, null, null);
-        hasServiceWithTags(serviceList, noTagsService, domainName, tagKey, updatedTagValues, 3);
-        hasServiceWithTags(serviceList, withTagsService, domainName, tagKey, multipleTags.get(tagKey).getList(), 3);
-        hasServiceWithTags(serviceList, withTagsService, domainName, "tagKey2", null, 0);
-
-        // get policies with exact tag value without unactive versions.
-        serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.FALSE, Boolean.FALSE, tagKey, "val3");
-        hasServiceWithTags(serviceList, noTagsService, domainName, tagKey, tagValues, 3);
-        hasServiceWithTags(serviceList, withTagsService, domainName, tagKey, newTagValues, 3);
-        assertEquals(serviceList.getList().size(), 2);
-
-        // get policies with no existing tag value
-        serviceList = zmsImpl.getServiceIdentities(ctx, domainName, Boolean.TRUE, Boolean.TRUE, tagKey, "val10");
-        assertEquals(serviceList.getList().size(), 0);
     }
 }
